Add files via upload
I just uploaded the FET extractor file for single reference. In the code, if your blank is not "h2o", change it to whatever your blank is, be sure to change it in your file name as well. This file is for only one reference (your background is stable). Let me know if that is not working for you. Feel free to talk to me if you need more modification. Enjoy your new program. :)
---ffffffrank